Add files via upload
The project aims to allow users to access and analyze data from multiple YouTube channels. The project utilizes SQL, MongoDB, and Streamlit to create a user-friendly application that allows users to retrieve, store, and query YouTube channel and video data.

The project consists of the following components:

YouTube Data API: The YouTube Data API is used to extract data about videos, channels, and users.
MongoDB: MongoDB is used to store the data in a database.
Streamlit: Streamlit is used to create a user-friendly dashboard that allows users to explore the data.

The project works as follows:

There are already predefined 10 channel_ids are stored in it we have to select one of them to get the details from YouTube and to see the details in MongoDB format
The YouTube Data API is used to extract data about the channel, such as the channel name, subscribers, total video count, playlist ID, video ID, likes, dislikes, and comments.
The data is stored in a MongoDB database.
The Streamlit dashboard is used to visualize the data. The dashboard allows users to filter, sort, and plot the data.

The project can be used to gain insights into user behavior, video popularity, and other trends. For example, the project can be used to answer questions such as:
Query Output need to displayed as table in Streamlit Application:

What are the names of all the videos and their corresponding channels?
Which channels have the most number of videos, and how many videos do they have?
What are the top 10 most viewed videos and their respective channels?
How many comments were made on each video, and what are their corresponding video names?
Which videos have the highest number of likes, and what are their corresponding channel names?
What is the total number of likes and dislikes for each video, and what are their corresponding video names?
What is the total number of views for each channel, and what are their corresponding channel names?
What are the names of all the channels that have published videos in the year 2022?
What is the average duration of all videos in each channel, and what are their corresponding channel names?
Which videos have the highest number of comments, and what are their corresponding channel names?

Here are some of the benefits of using this project:

It is a powerful tool for analyzing statistical data of YouTube data.
It is easy to use and can be used by anyone with basic knowledge of SQL and MongoDB and Streamlit.
It is open source, so you can modify it to fit your specific needs.